refactor(ai): M0 — split trade_insights_ai.py into 4-module package
The single-file module had grown to 2285 lines from accumulated v5/v5.1/v5.2
work — well past the project's 1000-line "stop and propose a split" budget.
Splitting now before v5.3 lands keeps each subsequent milestone working in
a focused module instead of growing the monolith further.

Layout (each module under 1000 lines):

  reports/trade_insights_ai/
    __init__.py        re-exports public surface (callers unchanged)
    prompt_text.py     MARKET_INTELLIGENCE_PROMPT + vocabulary constants
    analysis_input.py  build payload, hash, prompt, JSON Schema
    validators.py      deterministic HARD validators
    markdown.py        audit-view markdown renderer

Public import path stays stable — `from uw_scan.reports.trade_insights_ai
import X` keeps working for every X the router, worker, lenient module,
and tests use. The pre-existing circular dependency on
trade_insights_ai_lenient is handled the same way the single file did:
a late module-level re-export in __init__.py plus a function-local
import inside validate_trade_insights_ai_outcome. This is functionally
equivalent to the pre-split module — both versions only resolve when
trade_insights_ai is imported first (lenient-first import was already
broken before this refactor).

Verification:
  - uv run pytest tests/ -x -q --ignore=tests/integration
    643 passed, 3 skipped (no regressions)
  - Full public-surface import smoke (PROMPT_VERSION, all builders,
    validators, _iso_z, _to_decimal, _coerce_claude_outcome_dict) green.

Includes the v5.3 + outcome ledger implementation plan in
docs/superpowers/plans/.
