+++ src/uw_scan/reports/trade_insights_ai/markdown.py
@@ -1,0 +1,106 @@
+"""Markdown rendering for Trade Insights AI outcomes (audit view)."""
+from __future__ import annotations
+
+from uw_scan.models import TradeInsightAiOutcome
+
+from .analysis_input import _iso_z
+
+def render_trade_insights_ai_markdown(outcome: TradeInsightAiOutcome) -> str:
+    """Render compact Markdown from validated structured output."""
+
+    # `headline.score` is repurposed as a 0-100 confidence percentage. We
+    # render it as "Confidence: N/100", and suppress the line when the model
+    # left it blank (0/0) — drives the cosmetic v3 follow-up.
+    confidence_line: list[str] = []
+    if outcome.headline.score or outcome.headline.score_scale:
+        confidence_line.append(
+            f"Confidence: {outcome.headline.score}/{outcome.headline.score_scale}"
+        )
+    lines: list[str] = [
+        f"# {outcome.ticker} - {outcome.headline.stance_label}",
+        outcome.headline.title,
+        "",
+        f"Produced: {_iso_z(outcome.analysis_produced_at)}",
+        *confidence_line,
+        f"Conviction: {outcome.headline.conviction} - {outcome.headline.conviction_label}",
+        f"Top reason: {outcome.headline.top_reason}",
+        f"Primary risk: {outcome.headline.primary_risk}",
+        f"Watch: {outcome.headline.watch_trigger}",
+        "",
+        "## Metrics",
+    ]
+    for card in outcome.metric_cards:
+        note = f" - {card.note}" if card.note else ""
+        lines.append(f"- {card.label}: {card.value}{note}")
+
+    lines.extend(["", "## Scenarios"])
+    for card in outcome.scenario_cards:
+        lines.append(f"- {card.case}: {card.title} - {card.description}")
+
+    lines.extend(["", "## Sections"])
+    for section in (
+        outcome.section_cards.market_structure,
+        outcome.section_cards.volatility,
+        outcome.section_cards.flow_positioning,
+    ):
+        score = (
+            f" ({section.score}/{section.max_score})"
+            if section.score is not None and section.max_score is not None
+            else ""
+        )
+        lines.append(f"### {section.title}{score}")
+        lines.append(section.summary)
+        for highlight in section.highlights:
+            note = f" - {highlight.note}" if highlight.note else ""
+            lines.append(f"- {highlight.label}: {highlight.value}{note}")
+        for level in section.levels:
+            note = f" - {level.note}" if level.note else ""
+            lines.append(f"- {level.kind}: {level.price} {level.value}{note}")
+
+    if outcome.vrp_assessment is not None:
+        lines.extend(["", f"## {outcome.vrp_assessment.title}"])
+        lines.append(outcome.vrp_assessment.summary)
+        for metric in outcome.vrp_assessment.metrics:
+            lines.append(f"- {metric.label}: {metric.value}")
+        lines.append(f"Reason: {outcome.vrp_assessment.reason}")
+
+    if outcome.preferred_expression is not None:
+        expression = outcome.preferred_expression
+        lines.extend(["", f"## {expression.title}"])
+        if expression.subtitle:
+            lines.append(expression.subtitle)
+        lines.append(f"Why: {expression.why}")
+        lines.append(f"Status: {expression.status_observed}")
+        lines.append(
+            f"Risk flags: {', '.join(expression.risk_flags_observed) or 'none'}"
+        )
+        for note in expression.management_notes:
+            lines.append(f"- {note}")
+
+    if outcome.conflicts:
+        lines.extend(["", "## Conflicts"])
+        for item in outcome.conflicts:
+            lines.append(f"- {item.severity}: {item.description}")
+
+    if outcome.required_checks:
+        lines.extend(["", "## Required Checks"])
+        for item in outcome.required_checks:
+            blocker = "blocks sizing" if item.blocks_sizing else "informational"
+            # Strip trailing period from `check` so the title — reason join
+            # doesn't render as "...fires.: Liquidity must support..." (the
+            # model frequently terminates the check phrase with punctuation).
+            check_text = item.check.rstrip(". ").rstrip()
+            lines.append(f"- {check_text} — {item.reason} ({blocker})")
+
+    if outcome.rejected_ideas:
+        lines.extend(["", "## Rejected Ideas"])
+        for item in outcome.rejected_ideas:
+            lines.append(f"- {item.idea_id} {item.structure}: {item.reason}")
+
+    if outcome.missing_data:
+        lines.extend(["", "## Missing Data"])
+        for item in outcome.missing_data:
+            lines.append(f"- {item}")
+
+    lines.extend(["", outcome.rendering.disclaimer])
+    return "\n".join(lines)
